% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{createGithub}
\alias{createGithub}
\title{Create a project GitHub repository}
\usage{
createGithub(project_name, path, private = TRUE)
}
\arguments{
\item{project_name}{project name/title}

\item{path}{target file path for project directory}

\item{private}{TRUE/FALSE. Should the GitHub repository be private? Evaluated only if arguments \code{git} and \code{github} are TRUE.}
}
\description{
Create a GitHub repository for the project.
}
