% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker}
\alias{docker}
\title{Add docker infrastructure}
\usage{
docker(project_name, path)
}
\arguments{
\item{project_name}{project name/title}

\item{path}{target file path for project directory}
}
\description{
Add docker infrastructure to a project directory.
}
\examples{
projectSkeleton('test project',path = tempdir(),type = 'report')
docker('test project',path = tempdir())
}
