% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Add project outputs}
\usage{
output(project_name, project_directory, type = projectTypes())
}
\arguments{
\item{project_name}{project name/title}

\item{project_directory}{the project directory file path}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}
}
\description{
Add project communication outputs to a project directory.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
output('test project',paste0(tempdir(),'/test_project'),type = 'report')
}
}
