% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{projectDirectory}
\alias{projectDirectory}
\title{Project directory}
\usage{
projectDirectory(project_name, path)
}
\arguments{
\item{project_name}{the project name/title}

\item{path}{target file path for project directory}
}
\description{
Convert a project name to a directory path.
}
\examples{
projectDirectory('test project',getwd())
}
