% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{projectSkeleton}
\alias{projectSkeleton}
\title{Project skeleton}
\usage{
projectSkeleton(project_directory, force = FALSE)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{force}{force project creation if project directory already exists}
}
\description{
Initialise the basic project directories and files.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
}
}
