% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{readme}
\alias{readme}
\title{Project README}
\usage{
readme(project_name, path, type)
}
\arguments{
\item{project_name}{project name/title}

\item{path}{target file path for project directory}

\item{type}{project type}
}
\description{
Export a README to a project directory
}
\examples{
readme('test project',getwd(),'report')
}
