% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprofile.R
\name{Rprofile}
\alias{Rprofile}
\title{Add project .Rprofile}
\usage{
Rprofile(
  project_directory,
  packages = c("magrittr", "targets", "conflicted"),
  install = FALSE,
  renv = TRUE
)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{packages}{character vector R packages to load}

\item{install}{value to set install argument of `pacman` package loading functions}

\item{renv}{add infrastructure for renv package environment management}
}
\description{
Add a template .Rprofile file to a project directory.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
Rprofile(paste0(tempdir(),'/test_project'))
}
}
