% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targets}
\alias{targets}
\title{Add targets infrastructure}
\usage{
targets(project_directory, type = projectTypes())
}
\arguments{
\item{project_directory}{the project directory file path}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}
}
\description{
Add targets infrastructure to a project directory
}
\examples{
projectSkeleton('test project',path = tempdir(),type = 'report')
targets(paste0(tempdir(),'/test_project'),type = 'report') 
}
