% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targetsRun}
\alias{targetsRun}
\title{Add a targets run script to a project directory}
\usage{
targetsRun(project_directory)
}
\description{
Add a run.R script to a project directory that can be used to trigger building a targets workflow.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
targetsRun(paste0(tempdir(),'/test_project'))
}
}
