% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{project templates}
\usage{
template(
  project_name,
  path = ".",
  type = c("report", "manuscript", "presentation"),
  start = TRUE
)
}
\arguments{
\item{project_name}{project name/title}

\item{path}{target file path for project directory}

\item{type}{project type}

\item{start}{TRUE/FALSE should the project be automatically activated}
}
\description{
Generate a project template directory.
}
\examples{
\dontrun{
template('A new project',type = 'report',start = FALSE)
}
}
