% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{createGit}
\alias{createGit}
\title{Add git infrastructure}
\usage{
createGit(project_directory, type)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}
}
\description{
Add git infrastructure to a project directory.
}
\examples{
projectSkeleton('test project',path = tempdir(),type = 'report')
createGit(paste0(tempdir(),'/test_project'),type = 'report') 
}
