% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker}
\alias{docker}
\title{Add docker infrastructure}
\usage{
docker(project_name, type = "report", path, renv = TRUE)
}
\arguments{
\item{project_name}{project name/title}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}

\item{path}{target file path for project directory}

\item{renv}{add infrastructure for renv package environment management}
}
\description{
Add docker infrastructure to a project directory.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
docker('test project',path = tempdir())
}
}
