% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{readme}
\alias{readme}
\title{Project README}
\usage{
readme(project_name, path, type = projectTypes(), renv = TRUE)
}
\arguments{
\item{project_name}{project name/title}

\item{path}{target file path for project directory}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}

\item{renv}{add text for renv support}
}
\description{
Export a README to a project directory
}
\examples{
\dontrun{
readme('test project',getwd(),'report')
}
}
