% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renvInitialise}
\alias{renvInitialise}
\title{Initialise renv infrastructure}
\usage{
renvInitialise(
  project_directory,
  bioconductor = FALSE,
  dependencies = character(),
  sandbox = FALSE
)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{bioconductor}{The version of bioconductor to use with the project. Set to TRUE to use the default version of Bioconductor.}

\item{dependencies}{additional package dependencies to install into the cache upon initialisation}

\item{sandbox}{Should sandboxing be enabled for \code{renv}? See the \code{renv} \href{https://rstudio.github.io/renv/reference/config.html}{config documentation} for more information on sandboxing in \code{renv}.}
}
\description{
Initialise renv infrastructure in a project directory.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
renvInitialise(paste0(tempdir(),'/test_project'))
}
}
