% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renvInitialise}
\alias{renvInitialise}
\title{Initialise renv infrastructure}
\usage{
renvInitialise(project_directory, bioconductor = FALSE, github = character())
}
\arguments{
\item{project_directory}{the project directory file path}

\item{bioconductor}{The version of bioconductor to use with the project. Set to TRUE to use the default version of Bioconductor.}

\item{github}{GitHub packages to install into the cache upon initialisation}
}
\description{
Initialise renv infrastructure in a project directory.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
renvInitialise(paste0(tempdir(),'/test_project'))
}
}
