% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renvInitialise}
\alias{renvInitialise}
\title{Initialise renv infrastructure}
\usage{
renvInitialise(project_directory, rebuild = FALSE)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{rebuild}{Force packages to be rebuilt, thereby bypassing any installed versions of the package available in the cache? This can either be a boolean (indicating that all installed packages should be rebuilt), or a vector of package names indicating which packages should be rebuilt.}
}
\description{
Initialise renv infrastructure in a project directory.
}
\examples{
projectSkeleton('test project',path = tempdir(),type = 'report')
renvInitialise(paste0(tempdir(),'/test_project'))
}
