% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprofile.R
\name{Rprofile}
\alias{Rprofile}
\title{Add project .Rprofile}
\usage{
Rprofile(project_directory, renv = TRUE)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{renv}{add infrastructure for renv package environment management}
}
\description{
Add a template .Rprofile file to a project directory.
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
Rprofile(paste0(tempdir(),'/test_project'))
}
}
