% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targets}
\alias{targets}
\title{Add targets infrastructure}
\usage{
targets(project_directory, type = projectTypes(), ...)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}

\item{...}{arguments to pass to \code{targetsRun()}}
}
\description{
Add targets infrastructure to a project directory
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
targets(paste0(tempdir(),'/test_project'),type = 'report') 
}
}
