% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targets}
\alias{targets}
\title{Add targets infrastructure}
\usage{
targets(
  project_directory,
  type = projectTypes(),
  targets_config = list(main = list(reporter_make = "timestamp_positives")),
  renv = TRUE
)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}

\item{targets_config}{a list containing the targets configuration to be written to YAML format}

\item{renv}{interface the use of an renv package cache}
}
\description{
Add targets infrastructure to a project directory
}
\examples{
\dontrun{
projectSkeleton(paste0(tempdir(),'/test_project'))
targets(paste0(tempdir(),'/test_project'),type = 'report') 
}
}
