% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{project templates}
\usage{
template(
  project_name,
  path = ".",
  type = c("report", "manuscript", "presentation"),
  git = TRUE,
  github = TRUE,
  private = TRUE,
  start = TRUE
)
}
\arguments{
\item{project_name}{project name/title}

\item{path}{target file path for project directory}

\item{type}{project type}

\item{git}{TRUE/FALSE. Initialise a Git repository?}

\item{github}{TRUE/FALSE. Create a GitHub repository?}

\item{private}{TRUE/FALSE. Should the GitHub repository be private. Ignored if argument \code{github} is FALSE.}

\item{start}{TRUE/FALSE. Should the project be automatically activated?}
}
\description{
Generate a project template directory.
}
\examples{
\dontrun{
template('A new project',type = 'report',start = FALSE)
}
}
