% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\title{Add R package settings}
\usage{
utils(project_directory = ".")
}
\arguments{
\item{project_directory}{the project directory file path}
}
\description{
Add project utilities to a project directory such as package settings.
}
\examples{
\dontrun{
utils(paste0(tempdir(),'/test_project'))
}
}
