% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\title{Add R package loading utilities}
\usage{
utils(
  project_directory = ".",
  cran = character(),
  bioc = character(),
  github = character(),
  install = FALSE
)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{cran}{character vector of cran R package dependencies}

\item{bioc}{character vector of bioconductor R package dependencies}

\item{github}{character vector of GitHub R package dependencies in the form `repository/package_name`}

\item{install}{value to set install argument of `pacman` package loading functions}
}
\description{
Add package utilites to a project directory.
}
\examples{
\dontrun{
utils(paste0(tempdir(),'/test_project'),cran = 'tidyverse')
}
}
