% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\title{Add package utilities}
\usage{
utils(project_directory, type)
}
\arguments{
\item{project_directory}{the project directory file path}

\item{type}{project type. Should be one returned by \code{projectTypes()}.}
}
\description{
Add package utilites to a project directory.
}
\examples{
projectSkeleton('test project',path = tempdir(),type = 'report')
utils(paste0(tempdir(),'/test_project'),type = 'report') 
}
